requirements.txt
Django>=3.0.7
django-debug-toolbar==2.2
perceval==0.12.24
oauthlib==3.1.0
requests-oauthlib==1.3.
django-notifications-hq==1.6.0
django-crispy-forms==1.9.1
django-totd==1.0.2
django-imagekit==4.0.2
django-imagekit-cropper==1.16
Pillow>=6.2.2
djangorestframework==3.11.1
